update: система прогрессии

+++ WebApi.Infrastructure/Models/Storage/Progress.cs
@@ -9,5 +9,5 @@ public class Progress : BaseEntity
     public Theme Theme { get; set; } = null!;
     
     public int Level { get; set; }
-    public int AmountToLevelUp { get; set; }
+    public double AmountToLevelUp { get; set; }
 }
